+++ packages/ui/src/modal/buy/BuyModalRenderer.tsx
@@ -5,6 +5,7 @@ import React, {
   useCallback,
   ReactNode,
   useMemo,
+  useContext,
 } from 'react'
 import {
   useTokens,
@@ -28,6 +29,7 @@ import * as allChains from 'viem/chains'
 import usePaymentTokens, {
   EnhancedCurrency,
 } from '../../hooks/usePaymentTokens'
+import { ProviderOptionsContext } from '../../ReservoirKitProvider'
 
 type Item = Parameters<ReservoirClientActions['buyToken']>['0']['items'][0]
 
@@ -92,6 +94,7 @@ type Props = {
   open: boolean
   tokenId?: string
   chainId?: number
+  defaultQuantity?: number
   collectionId?: string
   orderId?: string
   feesOnTopBps?: string[] | null
@@ -110,6 +113,7 @@ export const BuyModalRenderer: FC<Props> = ({
   collectionId,
   orderId,
   feesOnTopBps,
+  defaultQuantity,
   feesOnTopUsd,
   normalizeRoyalties,
   onConnectWallet,
@@ -133,6 +137,7 @@ export const BuyModalRenderer: FC<Props> = ({
 
   const client = useReservoirClient()
   const currentChain = client?.currentChain()
+  const providerOptionsContext = useContext(ProviderOptionsContext)
 
   const rendererChain = chainId
     ? client?.chains.find(({ id }) => id === chainId) || currentChain
@@ -168,7 +173,9 @@ export const BuyModalRenderer: FC<Props> = ({
     address as Address,
     _paymentCurrency ?? chainCurrency,
     totalIncludingFees,
-    rendererChain?.id
+    rendererChain?.id,
+    false,
+    true
   )
 
   const paymentCurrency = paymentTokens?.find(
@@ -330,37 +337,83 @@ export const BuyModalRenderer: FC<Props> = ({
   ])
 
   useEffect(() => {
-    fetchPath()
-  }, [fetchPath])
+    // ensure the tokens api has been fetched first
+    if (token) {
+      fetchPath()
+    }
+  }, [fetchPath, token])
+
+  const getCurrencyDetails = useCallback(() => {
+    if (orderId) {
+      return {
+        currency: listing?.price?.currency?.contract,
+        decimals: listing?.price?.currency?.decimals,
+        symbol: listing?.price?.currency?.symbol,
+        chainId: rendererChain?.id || 1,
+      }
+    } else if (is1155) {
+      return {
+        currency: path?.[0]?.currency,
+        decimals: path?.[0]?.currencyDecimals,
+        symbol: path?.[0]?.currencySymbol,
+        chainId: rendererChain?.id || 1,
+      }
+    } else {
+      return {
+        currency: token?.market?.floorAsk?.price?.currency?.contract,
+        decimals: token?.market?.floorAsk?.price?.currency?.decimals,
+        symbol: token?.market?.floorAsk?.price?.currency?.symbol,
+        chainId: rendererChain?.id || 1,
+      }
+    }
+  }, [orderId, is1155, listing, path, token, rendererChain])
 
   useEffect(() => {
-    if (!paymentTokens[0] || paymentCurrency) {
+    if (
+      !paymentTokens[0] ||
+      paymentCurrency ||
+      !token ||
+      (orderId && !listing) ||
+      (is1155 && !path)
+    ) {
       return
     }
 
-    let selectedCurrency
+    const { currency, decimals, symbol, chainId } = getCurrencyDetails()
 
-    if (orderId) {
-      selectedCurrency =
-        paymentTokens.find(
-          (token) =>
-            token.address === listing?.price?.currency?.contract?.toLowerCase()
-        ) || paymentTokens[0]
-    } else if (is1155) {
-      selectedCurrency =
-        paymentTokens.find(
-          (token) => token.address === path?.[0].currency?.toLowerCase()
-        ) || paymentTokens[0]
+    // Determine whether to include the listing currency unconditionally
+    const includeListingCurrency =
+      providerOptionsContext.alwaysIncludeListingCurrency !== false
+
+    let selectedCurrency
+    if (includeListingCurrency && currency) {
+      selectedCurrency = {
+        address: currency.toLowerCase() as Address,
+        decimals: decimals || 18,
+        symbol: symbol || '',
+        chainId: chainId,
+      }
     } else {
-      selectedCurrency =
-        paymentTokens.find(
-          (paymentToken) =>
-            paymentToken.address ===
-            token?.market?.floorAsk?.price?.currency?.contract?.toLowerCase()
-        ) || paymentTokens[0]
+      selectedCurrency = paymentTokens.find(
+        (token) => token.address === currency?.toLowerCase()
+      )
+
+      if (!selectedCurrency) {
+        selectedCurrency = paymentTokens[0]
+      }
     }
+
     setPaymentCurrency(selectedCurrency)
-  }, [paymentTokens, chainCurrency])
+  }, [
+    paymentTokens,
+    paymentCurrency,
+    orderId,
+    is1155,
+    listing,
+    path,
+    token,
+    providerOptionsContext.alwaysIncludeListingCurrency,
+  ])
 
   const buyToken = useCallback(async () => {
     if (!wallet) {
@@ -715,8 +768,17 @@ export const BuyModalRenderer: FC<Props> = ({
       setQuantity(1)
       setPath(undefined)
       setPaymentCurrency(undefined)
+      setToken(undefined)
+    } else {
+      setQuantity(defaultQuantity || 1)
     }
   }, [open])
+
+  useEffect(() => {
+    if (quantityRemaining > 0 && quantity > quantityRemaining) {
+      setQuantity(quantityRemaining)
+    }
+  }, [quantityRemaining, quantity])
 
   return (
     <>