+++ packages/ui/src/primitives/ErrorWell.tsx
@@ -25,6 +25,8 @@ export default function ErrorWell({ error, css }: Props) {
       message = error.message
     } else if (error.name === 'TransactionTimeoutError') {
       txHash = (error as TransactionTimeoutError).txHash
+    } else if (error.name === 'CrossChainTransactionError') {
+      message = error.message
     } else if (errorStatus >= 400 && errorStatus < 500) {
       message = error.message
     }